学習結果のテスト時に評価値をi32に変換するのをやめて、f32のまま出力するようにした。

+++ src/learning.rs
@@ -514,9 +514,9 @@ impl<NN> Learnener<NN>
 
 				if success {
 					successed += 1;
-					println!("評価値{} 正解!",(score * (1 << 29) as f32) as i32);
+					println!("評価値{} 正解!",score * (1 << 29) as f32);
 				} else {
-					println!("評価値{} 不正解...",(score * (1 << 29) as f32) as i32);
+					println!("評価値{} 不正解...",score * (1 << 29) as f32);
 				}
 
 				count += 1;
@@ -907,9 +907,9 @@ impl<NN> Learnener<NN>
 
 				if success {
 					successed += 1;
-					println!("評価値{} 正解!",(score * (1 << 29) as f32) as i32);
+					println!("評価値{} 正解!",score * (1 << 29) as f32);
 				} else {
-					println!("評価値{} 不正解...",(score * (1 << 29) as f32) as i32);
+					println!("評価値{} 不正解...",score * (1 << 29) as f32);
 				}
 
 				count += 1;
